+++ lib/generators/anycable/setup/setup_generator.rb
@@ -28,15 +28,7 @@ module AnyCableRailsGenerators
       default: "latest"
     class_option :postgres_signalling,
       type: :boolean,
-      desc: "Install the Postgres signalling contract and use Postgres for broadcasting and pub/sub"
-
-    def self.next_migration_number(dirname)
-      if ActiveRecord.timestamped_migrations
-        Time.now.utc.strftime("%Y%m%d%H%M%S")
-      else
-        "%.3d" % (current_migration_number(dirname) + 1)
-      end
-    end
+      desc: "Use Postgres for broadcasting and pub/sub"
 
     def welcome
       say ""
@@ -132,13 +124,6 @@ module AnyCableRailsGenerators
       template "anycable.toml"
 
       update_cable_yml
-    end
-
-    def postgres_signalling_contract
-      return unless postgres?
-
-      active_record_migration_support!
-      migration_template "db/migrate/create_anycable_postgres_signalling.rb", "db/migrate/create_anycable_postgres_signalling.rb"
     end
 
     def rubocop_compatibility
@@ -266,20 +251,6 @@ module AnyCableRailsGenerators
 
     def postgres?
       !!options[:postgres_signalling]
-    end
-
-    def active_record_migration_support!
-      require "rails/generators/active_record/migration"
-
-      self.class.include ActiveRecord::Generators::Migration unless self.class < ActiveRecord::Generators::Migration
-    rescue LoadError
-      raise Thor::Error, "ActiveRecord is required to generate the Postgres signalling migration"
-    end
-
-    def migration_version
-      "[#{ActiveRecord::Migration.current_version}]"
-    rescue NoMethodError
-      "[7.0]"
     end
 
     def webpacker?