+++ spec/generators/setup/setup_generator_spec.rb
@@ -23,7 +23,6 @@ describe AnyCableRailsGenerators::SetupGenerator, type: :generator do
   let(:cable_yml) { YAML.load_file(File.join(destination_root, "config/cable.yml"), aliases: true).deep_symbolize_keys }
   let(:anycable_yml) { YAML.load_file(File.join(destination_root, "config/anycable.yml"), aliases: true).deep_symbolize_keys }
   let(:anycable_toml) { TOML.load_file(File.join(destination_root, "anycable.toml")).deep_symbolize_keys }
-  let(:postgres_migration) { Dir[File.join(destination_root, "db/migrate/*_create_anycable_postgres_signalling.rb")].first }
 
   # Base options=skip everything, the generator only adds/updates config files
   let(:base_opts) { {rpc: "none", development: "skip"} }
@@ -131,23 +130,19 @@ describe AnyCableRailsGenerators::SetupGenerator, type: :generator do
       subject
 
       expect(anycable_yml.dig(:development, :broadcast_adapter)).to eq "postgres"
-      expect(anycable_yml.dig(:development, :postgres_broadcasts_table)).to eq "anycable_broadcasts"
-      expect(anycable_yml.dig(:development, :postgres_contract_table)).to eq "anycable_contracts"
+      expect(anycable_yml.dig(:development, :postgres_broadcasts_table)).to be_nil
+      expect(anycable_yml.dig(:development, :postgres_contract_table)).to be_nil
 
       expect(anycable_toml[:broadcast_adapters]).to eq(["http", "postgres"])
       expect(anycable_toml[:pubsub_adapter]).to eq "postgres"
-      expect(anycable_toml.dig(:postgres, :notify_channel)).to eq "anycable_signals"
+      expect(anycable_toml.dig(:postgres, :broadcast_notify_channel)).to eq "anycable_broadcasts"
+      expect(anycable_toml.dig(:postgres, :pubsub_notify_channel)).to eq "anycable_pubsub"
       expect(anycable_toml.dig(:postgres, :broadcasts_table)).to eq "anycable_broadcasts"
       expect(anycable_toml.dig(:postgres, :pubsub_table)).to eq "anycable_pubsub"
-      expect(anycable_toml.dig(:postgres, :contract_table)).to eq "anycable_contracts"
-      expect(anycable_toml.dig(:postgres, :validate_contract)).to eq true
+      expect(anycable_toml.dig(:postgres, :ensure_schema)).to eq true
+      expect(anycable_toml.dig(:postgres, :exhausted_broadcast_policy)).to eq "skip"
 
-      expect(postgres_migration).not_to be_nil
-      migration = File.read(postgres_migration)
-      expect(migration).to include("CREATE TABLE anycable_broadcasts")
-      expect(migration).to include("CREATE TABLE anycable_pubsub")
-      expect(migration).to include("CREATE TRIGGER anycable_broadcasts_notify_insert")
-      expect(migration).to include("CREATE TRIGGER anycable_pubsub_notify_insert")
+      expect(Dir[File.join(destination_root, "db/migrate/*_create_anycable_postgres_signalling.rb")]).to be_empty
     end
   end